Split out `complete_macro_in_item_position`
